Separated pieces of the main.py into their own files for better organization.
